PKG: do not apply lint rule that has false positives
Biomejs noFocusedTests is currently marking the fit() method
call on the xterm.js FitAddon as suspicious. This rule is only
useful for tests written with Jest [1], so clearly this is a
false positive and should be ignored. Because I am not using
the relevant feature of Jest much, easiest to ignore this rule
at the configuration file level.

[1] https://github.com/jest-community/eslint-plugin-jest/blob/63734225a67822849d28e405e2ec3b70f8c76302/docs/rules/no-focused-tests.md